+++ website/docusaurus.config.js
@@ -201,6 +201,12 @@ async function createConfig() {
     defer: true,
   });
 
+  config.scripts.push({
+    src: `${config.baseUrl}z3-built.js`,
+    type: 'text/javascript',
+    defer: true,
+  });
+
 
   return config;
 }
